+++ Bittrex/JsonData/CurrencyJson.cs
@@ -8,7 +8,7 @@ namespace Bittrex.JsonData
         public string Currency { get; set; }
         public string CurrencyLong { get; set; }
         public int MinConfirmation { get; set; }
-        public double TxFee { get; set; }
+        public decimal TxFee { get; set; }
         public bool IsActive { get; set; }
         public string CoinType { get; set; }
         public object BaseAddress { get; set; }
