+++ Bittrex/JsonData/OrderBook/Sell.cs
@@ -5,7 +5,7 @@ namespace Bittrex.JsonData
     [JsonObject(Title = "Sell")]
     public class SellJson
     {
-        public double Quantity { get; set; }
-        public double Rate { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal Rate { get; set; }
     }
 }
