+++ Common/TradeManager.cs
@@ -7,13 +7,13 @@ namespace Common
     public class TradeManager : ITradeManager
     {
         public IRealTimeData RealTimeData { get; }
-        public bool SetBuyOrder(int marketId, double quality, double rate)
+        public bool SetBuyOrder(int marketId, decimal quality, decimal rate)
         {
             Console.WriteLine($"{marketId}-{quality}-{rate}");
             return true;
         }
 
-        public bool SetSellOrder(int marketId, double quality, double rate)
+        public bool SetSellOrder(int marketId, decimal quality, decimal rate)
         {
             //todo
             return true;
