+++ DAL/OrderPoco.cs
@@ -29,17 +29,17 @@ namespace DAL
         /// <summary>
         /// Ставка (в основной валюте)
         /// </summary>
-        public double Rate { get; set; }
+        public decimal Rate { get; set; }
         
         /// <summary>
         /// Объём (в основной валюте)
         /// </summary>
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
         
         /// <summary>
         /// ???
         /// </summary>
-        public double BaseVolume { get; set; }
+        public decimal BaseVolume { get; set; }
         
         /// <summary>
         /// Время, когда была выставлена ставка
